requirements.txt
pyOpenSSL==16.2.0
boto3==1.3.1
PyYAML==3.11
pycrypto==2.6.1
certbot==0.26.1
josepy==1.1.0